+++ R/animateSpectra.R
@@ -23,23 +23,16 @@
 #'   species in the system is plotted as well. Default is FALSE.
 #'
 #' @return A plotly object
-#' @export
 #' @family plotting functions
-animateSpectra <- function(sim,
-                           species = NULL,
-                           resources = NULL,
-                           time_range,
-                           wlim = c(NA, NA),
-                           ylim = c(NA, NA),
-                           power = 1,
-                           total = FALSE) {
-    if (is.null(getComponent(getParams(sim), "MR"))) {
-        return(mizer::animateSpectra(sim = sim, species = species,
-                                     time_range = time_range,
-                                     wlim = wlim, ylim = ylim,
-                                     power = power, total = total,
-                                     resource = TRUE))
-    }
+#' @export
+animateSpectra.mizerMRSim <- function(sim,
+                                      species = NULL,
+                                      resources = NULL,
+                                      time_range,
+                                      wlim = c(NA, NA),
+                                      ylim = c(NA, NA),
+                                      power = 1,
+                                      total = FALSE) {
     species <- valid_species_arg(sim, species)
     resources <- valid_resources_arg(sim, resources)
     if (missing(time_range)) {