+++ tests/testthat/test-setMultipleResources.R
@@ -5,16 +5,19 @@ test_that("Can reproduce single resource behaviour", {
                      dim = c(1, length(NS_params@w_full)))
     params <- setMultipleResources(NS_params, resource_params = rp,
                                    initial_resource = initial)
-    expect_identical(getEncounter(params), getEncounter(NS_params))
-    expect_identical(getResourceMort(params)[1, ], getResourceMort(NS_params))
+    expect_equal(getEncounter(params), getEncounter(NS_params),
+                 ignore_attr = TRUE)
+    expect_equal(getResourceMort(params)[1, ], getResourceMort(NS_params),
+                 ignore_attr = TRUE)
     sim <- project(params, t_max = 10, dt = 1)
     simo <- project(NS_params, t_max = 10, dt = 1)
     expect_equal(sim@n_other[[1, "MR"]][1, ], simo@n_pp[1, ])
     expect_equal(sim@n_other[[11, "MR"]][1, ], simo@n_pp[11, ])
 
     # test that extension field in metadata is set
-    expect_equal(getMetadata(params)$extensions[["mizerMR"]],
-                 "sizespectrum/mizerMR")
+    expect_true(is.na(getMetadata(params)$extensions[["mizerMR"]]))
+    expect_equal(params@rates_funcs$Encounter, "mizerEncounter")
+    expect_equal(params@rates_funcs$ResourceMort, "mizerResourceMort")
 })
 
 test_that("Works with two identical resources", {
@@ -29,8 +32,10 @@ test_that("Works with two identical resources", {
                                    initial_resource = initial)
     expect_identical(params@initial_n_other[["MR"]][1, ],
                      NS_params@initial_n_pp / 2)
-    expect_identical(getEncounter(params), getEncounter(NS_params))
-    expect_identical(getResourceMort(params)[1, ], getResourceMort(NS_params))
+    expect_equal(getEncounter(params), getEncounter(NS_params),
+                 ignore_attr = TRUE)
+    expect_equal(getResourceMort(params)[1, ], getResourceMort(NS_params),
+                 ignore_attr = TRUE)
     sim <- project(params, t_max = 10)
     simo <- project(NS_params, t_max = 10)
     expect_equal(sim@n_other[[1, "MR"]][1, ], simo@n_pp[1, ] / 2)
@@ -125,4 +130,33 @@ test_that("Work with non-MR objects", {
                      mizer::resource_capacity(NS_params))
     expect_identical(resource_rate(NS_params),
                      mizer::resource_rate(NS_params))
+})
+
+test_that("Multiple resources compose with outer encounter methods", {
+    projectEncounter.doubleSearch <- function(params, ...) {
+        params@search_vol <- 2 * params@search_vol
+        NextMethod()
+    }
+    assign("projectEncounter.doubleSearch", projectEncounter.doubleSearch,
+           envir = .GlobalEnv)
+    on.exit(rm(projectEncounter.doubleSearch, envir = .GlobalEnv),
+            add = TRUE)
+
+    rp <- as.data.frame(NS_params@resource_params)
+    rp$resource <- "main"
+    initial <- array(NS_params@initial_n_pp,
+                     dim = c(1, length(NS_params@w_full)))
+    params <- setMultipleResources(NS_params, resource_params = rp,
+                                   initial_resource = initial)
+    encounter <- getEncounter(params)
+
+    params@extensions <- c(doubleSearch = NA_character_, params@extensions)
+    mizer::registerExtensions(params@extensions)
+    params <- mizer::coerceToExtensionClass(params)
+
+    expect_equal(getEncounter(params), 2 * encounter, ignore_attr = TRUE)
+
+    old_extensions <- params@extensions
+    resource_rate(params) <- resource_rate(params)
+    expect_identical(params@extensions, old_extensions)
 })